Rename folder to pt
Though I understand there is significant difference between pt-pt to pt-br, @GuildBot and the pirates can't tell the difference between the version of Portuguese without guessing. (Even Google Translate does not know how to differentiate). We can look at using variants when it comes to output like Botswain, GUS and Eric but for the hails we only need one. The pt-br should be sufficient as most guilds are pt-br not pt-pt.